pyproject.toml
[project]
name = "petmatch-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "asyncio>=4.0.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "greenlet>=3.2.4",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "quart>=0.20.0",
    "quart-auth>=0.11.0",
    "quart-schema>=0.22.0",
    "ruff>=0.13.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
